// src/main/webapp/tds/web-app/app-js/modules/taskManager/model/move-event.model.ts
export interface IMoveEvent {
	actualCompletionTime?: string;
	actualStartTime?: string;
	apiActionBypass?: boolean;
	calcMethod?: string;
	dateCreated?: string;
	description?: string;
	estCompletionTime?: string;
	estStartTime?: string;
	id?: number;
	lastUpdated?: string;
	moveBundlesString?: string
	name?: string
	revisedCompletionTime?: string;
	runbookBridge1?: string
	runbookBridge2?: string;
	runbookRecipe?: string;
	runbookStatus?: string;
	runbookVersion?: number
	version?: number;
	videolink?: string;
}